+++ tests/Event/EndEventTest.php
@@ -3,8 +3,8 @@
 namespace Phlow\Tests\Event;
 
 use Phlow\Engine\Exchange;
-use Phlow\Handler\SingleConnectionHandler;
-use Phlow\Handler\UnmatchedConditionException;
+use Phlow\Processor\SingleConnectionProcessor;
+use Phlow\Processor\UnmatchedConditionException;
 use Phlow\Event\EndEvent;
 use PHPUnit\Framework\TestCase;
 
@@ -13,7 +13,7 @@ class EndEventTest extends TestCase
     public function testNext()
     {
         $task = new EndEvent();
-        $handler = new SingleConnectionHandler();
+        $handler = new SingleConnectionProcessor();
 
         $this->expectException(UnmatchedConditionException::class);
         $handler->handle($task, new Exchange());
